feat(stores.ts, SwList): 'time()' is stored in stores
