+++ src/StopwatchList.svelte
@@ -14,7 +14,7 @@
                 timestamp: new Date().getTime(),
                 seconds: 0,
                 started: false,
-                time: time(0),
+                time: stopwatches.time(0),
             };
         }));
     }
@@ -22,21 +22,6 @@
     let counting = 0;
     let doubleClick: string;
     let now: number;
-
-    /**
-     * convert seconds to HH:MM:SS
-     */
-    function time(seconds: number){
-        // convert to HH:MM:SS
-        let h: number | string = Math.floor(seconds / 3600);
-        let m: number | string = Math.floor(seconds / 60) % 60;
-        let s: number | string = seconds % 60;
-
-        if(h < 10) h = '0' + h;
-        if(m < 10) m = '0' + m;
-        if(s < 10) s = '0' + s;
-        return h + ':' + m + ':' + s;
-    }
 
     /**
      * returns number of seconds since start time to now
@@ -53,7 +38,7 @@
         now = new Date().getTime();
         stopwatches.update(sws => sws.map(sw => {
             if(sw.started){
-                sw.time = time(sw.seconds + delta(sw.timestamp));
+                sw.time = stopwatches.time(sw.seconds + delta(sw.timestamp));
             }
             return sw;
         }));
